requirements.txt
typer
rich
google-genai
pillow
python-dotenv
requests
